# hw04_prometheus/api/otus55-users-1.0.0-swagger.json
{
  "openapi" : "3.0.0",
  "info" : {
    "description" : "This is simple client API\n",
    "version" : "1.0.0",
    "title" : "User Service",
    "contact" : {
      "email" : "schetinnikov@gmail.com"
    }
  },
  "servers" : [ {
    "description" : "Example Server",
    "url" : "http://example.domain/api/v1/"
  }, {
    "description" : "SwaggerHub API Auto Mocking",
    "url" : "https://virtserver.swaggerhub.com/otus55/users/1.0.0"
  } ],
  "tags" : [ {
    "name" : "user",
    "description" : "Operations about user"
  } ],
  "paths" : {
    "/user" : {
      "post" : {
        "tags" : [ "user" ],
        "summary" : "Create user",
        "description" : "This can only be done by the logged in user.",
        "operationId" : "createUser",
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              },
              "examples" : {
                "sample-user" : {
                  "summary" : "Example",
                  "value" : {
                    "username" : "johndoe589",
                    "firstName" : "John",
                    "lastName" : "Doe",
                    "email" : "bestjohn@doe.com",
                    "phone" : "+71002003040"
                  }
                }
              }
            }
          },
          "description" : "Created user object",
          "required" : true
        }
      }
    },
    "/user/{userId}" : {
      "parameters" : [ {
        "name" : "userId",
        "in" : "path",
        "description" : "ID of user",
        "required" : true,
        "schema" : {
          "type" : "integer",
          "format" : "int64"
        }
      } ],
      "get" : {
        "tags" : [ "user" ],
        "description" : "Returns a user based on a single ID, if the user does not have access to the user",
        "operationId" : "find user by id",
        "responses" : {
          "200" : {
            "description" : "user response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            }
          },
          "default" : {
            "description" : "unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "user" ],
        "description" : "deletes a single user based on the ID supplied",
        "operationId" : "deleteUser",
        "responses" : {
          "204" : {
            "description" : "user deleted"
          },
          "default" : {
            "description" : "unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "user" ],
        "description" : "Update user with User ID supplied",
        "operationId" : "updateUser",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              },
              "examples" : {
                "sample-user" : {
                  "summary" : "Example",
                  "value" : {
                    "firstName" : "Julie",
                    "lastName" : "Doe",
                    "email" : "bestjohn@doe.com",
                    "phone" : "+71004242424"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "user updated"
          },
          "default" : {
            "description" : "unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "User" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "username" : {
            "type" : "string",
            "maxLength" : 256
          },
          "firstName" : {
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string",
            "format" : "email"
          },
          "phone" : {
            "type" : "string",
            "format" : "phone"
          }
        }
      },
      "Error" : {
        "type" : "object",
        "required" : [ "code", "message" ],
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          }
        }
      }
    },
    "requestBodies" : {
      "UserArray" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/User"
              }
            }
          }
        },
        "description" : "List of user object",
        "required" : true
      }
    }
  }
}